feat: interactive 'store add'
actually, this is mostly refactoring, to achieve that 'store add' uses the same
interface as the 'store edit'. so interactive flag came from that. but
still should work great. maybe some additional logic is needed to check
for the already existing images in the store to force users to use edit
on them instead.