fix('store add'): new-only
add only new images with add. if the user wants to edit something
already in the store, the user should use edit.

this functionality could be useful, for example, if the user wants to
edit the tags in the store of the wallpaper he knows is already stored,
but to support this functionality edit command in the future could just
support other ways of input: if the hash is not present, try to read it
like a path and get a hash.